requirements.txt
streamlit
openai
fpdf
python-docx
pdfminer.six
